Use an existing tibble as the basis for a spatial query

If you have a tibble such as those produced by create_lookup_table() - that is, there is a column of geographical ONS codes ending in 'cd' - simply use this table as the basis for retrieving the relevant boundaries.

Usage

add_geometry_to_table(
  tbl,
  points = FALSE,
  lookup = NULL,
  geo_code_field = NULL,
  crs = 4326,
  resolution = c("BGC", "BSC", "BUC", "BFC", "BFE")
)

Arguments

tbl

A tibble with a column for geographical codes. This function will use the lefthand-most column ending in 'cd' by default as the basis for retrieving boundary or point data

points

Whether to retrieve area centroids (TRUE) or area boundaries (FALSE, the default)

lookup

character. Lower level area code eg "lsoa", "wd", "lad". Equivalent to the lookup parameter in bounds().

geo_code_field

character The column name from tbl for which you wish to retrieve spatial data. By default, add_bounds_to_table() will use the lefthand-most column that ends in "cd"

crs

the Coordinate Reference System (CRS) code to use. 4326 by default.

resolution

character. See argument options for available options. How generalised should the boundary be, and whether coastal boundaries should adhere to the coastline or to the full territorial extent. BGC by default (G = Generalised (20m), C = limited to the coastline.) F indicates Full resolution; S indicates Super-generalised (200m); U indicates Ultra- generalised (500m) boundary resolution. Use E instead of C for full extent boundaries (e.g. BFE). Not all combinations are available.

Value

If successful, will return the initial table with an additional geometry column added. Duplicate rows will be removed.