Return boundary data at a specified level and area from the ONS OG API

Return boundary data at a specified level and area from the ONS OG API

Usage

bounds(
  lookup,
  within = NULL,
  within_names = NULL,
  within_codes = NULL,
  return_width = c("tidy", "full", "minimal"),
  lookup_year = NULL,
  within_year = NULL,
  country_filter = c("UK|GB|EW|EN|SC|WA", "UK", "GB", "EW", "EN", "SC", "WA"),
  option = NULL,
  crs = 4326,
  resolution = c("BGC", "BSC", "BUC", "BFC", "BFE")
)

Arguments

lookup

character. Lower level area code eg "lsoa", "wd", "lad". Equivalent to the lookup parameter in bounds().

within

character. Higher level area code eg "lad", "cty". Equivalent to the within parameter in bounds().

within_names, within_codes

character. In order to restrict data returned to a specific area, either within_names or within_codes must be provided. Otherwise all boundaries within the country at that level will be retrieved. Use place names eg "Essex" to restrict to a certain geographical area. Use ONS area codes eg "W02000103" likewise (this is useful with wards, where there are many that share identical names). A vector of multiple names or codes can be supplied.

return_width

character. How many of the possible columns in the returned table to keep. Options are "tidy", "full" or "minimal".

lookup_year

numeric or character. A specific year for data relating to parameter x, if needed. Defaults to NULL, which will return the most recent data.

within_year

numeric or character. A specific year for data relating to parameter y, if needed. Defaults to NULL, which will return the most recent data.

country_filter

character. Open Geography datasets are sometimes available just within certain countries. Specify a country code if you want your results restricted to a certain country only - eg "WA" for Wales, "EW" for England and Wales. By default returns all options.

option

numeric. Defaults to 1, which means that the URL will just be the first one from the list of possible services resulting from the level and year filters above. If this does not give you what you want, you can run the script again with a different option from the list.

crs

the Coordinate Reference System (CRS) code to use. 4326 by default.

resolution

character. See argument options for available options. How generalised should the boundary be, and whether coastal boundaries should adhere to the coastline or to the full territorial extent. BGC by default (G = Generalised (20m), C = limited to the coastline.) F indicates Full resolution; S indicates Super-generalised (200m); U indicates Ultra- generalised (500m) boundary resolution. Use E instead of C for full extent boundaries (e.g. BFE). Not all combinations are available.

Value

an sfc tibble (data frame with geometry)

Examples

bounds("msoa", "lad", "Swansea")
#> Simple feature collection with 30 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -4.333587 ymin: 51.53577 xmax: -3.842747 ymax: 51.77423
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>     msoa21cd    msoa21nm    msoa21nmw            msoa21hclnm
#> 1  W02000168 Swansea 001 Abertawe 001          Pontarddulais
#> 2  W02000169 Swansea 002 Abertawe 002         Clydach & Mawr
#> 3  W02000170 Swansea 003 Abertawe 003        Morriston North
#> 4  W02000171 Swansea 004 Abertawe 004 Llangyfelach & Tircoed
#> 5  W02000172 Swansea 005 Abertawe 005              Gorseinon
#> 6  W02000173 Swansea 006 Abertawe 006             Birchgrove
#> 7  W02000174 Swansea 007 Abertawe 007                Loughor
#> 8  W02000175 Swansea 008 Abertawe 008        Morriston South
#> 9  W02000176 Swansea 009 Abertawe 009            Mynydd-bach
#> 10 W02000177 Swansea 010 Abertawe 010             Llansamlet
#>               msoa21hclnmw   lad24cd lad24nm                       geometry
#> 1            Pontarddulais W06000011 Swansea MULTIPOLYGON (((-4.044836 5...
#> 2           Clydach a Mawr W06000011 Swansea MULTIPOLYGON (((-3.976283 5...
#> 3         Gogledd Treforys W06000011 Swansea MULTIPOLYGON (((-3.913253 5...
#> 4  Llangyfelach a Thircoed W06000011 Swansea MULTIPOLYGON (((-4.011018 5...
#> 5                Gorseinon W06000011 Swansea MULTIPOLYGON (((-4.062941 5...
#> 6                Gellifedw W06000011 Swansea MULTIPOLYGON (((-3.875069 5...
#> 7               Casllwchwr W06000011 Swansea MULTIPOLYGON (((-4.053356 5...
#> 8              De Treforys W06000011 Swansea MULTIPOLYGON (((-3.929819 5...
#> 9              Mynydd-bach W06000011 Swansea MULTIPOLYGON (((-3.933911 5...
#> 10              Llansamlet W06000011 Swansea MULTIPOLYGON (((-3.898853 5...
bounds("msoa", "lad", "Shepway", lookup_year = 2011, within_year = 2015)
#> Simple feature collection with 14 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0.7770266 ymin: 50.91095 xmax: 1.221137 ymax: 51.205
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>     msoa11cd    msoa11nm                          msoa11hclnm   lad15cd lad15nm
#> 1  E02005102 Shepway 001             Lyminge, Densole & Elham E07000112 Shepway
#> 2  E02005103 Shepway 002                             Hawkinge E07000112 Shepway
#> 3  E02005104 Shepway 003                Folkestone North East E07000112 Shepway
#> 4  E02005105 Shepway 004                     Folkestone Foord E07000112 Shepway
#> 5  E02005106 Shepway 005               Cheriton & Horn Street E07000112 Shepway
#> 6  E02005107 Shepway 006       Folkestone Morehall & Sandgate E07000112 Shepway
#> 7  E02005109 Shepway 008     Saltwood, Seabrook & Etchinghill E07000112 Shepway
#> 8  E02005110 Shepway 009                    Lympne & Palmarsh E07000112 Shepway
#> 9  E02005111 Shepway 010                                Hythe E07000112 Shepway
#> 10 E02005112 Shepway 011 Dymchurch, St Mary's Bay & Brookland E07000112 Shepway
#>                          geometry
#> 1  POLYGON ((1.089762 51.18973...
#> 2  POLYGON ((1.170835 51.13259...
#> 3  POLYGON ((1.184798 51.09985...
#> 4  POLYGON ((1.189394 51.08954...
#> 5  POLYGON ((1.135277 51.09801...
#> 6  POLYGON ((1.17032 51.0953, ...
#> 7  POLYGON ((1.111727 51.12941...
#> 8  POLYGON ((1.000461 51.1184,...
#> 9  POLYGON ((1.062613 51.07642...
#> 10 POLYGON ((0.9884286 51.0706...
bounds("rgn", country_filter = "EN", resolution = "BUC")
#> Simple feature collection with 9 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -6.360297 ymin: 49.88234 xmax: 1.763706 ymax: 55.81121
#> Geodetic CRS:  WGS 84
#>     rgn23cd                  rgn23nm                       geometry
#> 1 E12000001               North East MULTIPOLYGON (((-1.401114 5...
#> 2 E12000002               North West MULTIPOLYGON (((-3.242936 5...
#> 3 E12000003 Yorkshire and The Humber MULTIPOLYGON (((-0.5207704 ...
#> 4 E12000004            East Midlands MULTIPOLYGON (((-0.248476 5...
#> 5 E12000005            West Midlands MULTIPOLYGON (((-1.860821 5...
#> 6 E12000006          East of England MULTIPOLYGON (((0.6458286 5...
#> 7 E12000007                   London MULTIPOLYGON (((-0.01191868...
#> 8 E12000008               South East MULTIPOLYGON (((-1.55515 50...
#> 9 E12000009               South West MULTIPOLYGON (((-6.349051 4...
bounds("parish", "lad", "Isles of Scilly")
#> Simple feature collection with 5 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -6.418667 ymin: 49.86479 xmax: -6.244608 ymax: 49.98065
#> Geodetic CRS:  WGS 84
#>     par23cd     par23nm   lad23cd         lad23nm
#> 1 E04012729      Bryher E06000053 Isles of Scilly
#> 2 E04012730    St Agnes E06000053 Isles of Scilly
#> 3 E04012731 St Martin's E06000053 Isles of Scilly
#> 4 E04012732   St Mary's E06000053 Isles of Scilly
#> 5 E04012733      Tresco E06000053 Isles of Scilly
#>                         geometry
#> 1 MULTIPOLYGON (((-6.386119 4...
#> 2 MULTIPOLYGON (((-6.398976 4...
#> 3 MULTIPOLYGON (((-6.273013 4...
#> 4 MULTIPOLYGON (((-6.280232 4...
#> 5 MULTIPOLYGON (((-6.361791 4...
bounds("spr")
#> Simple feature collection with 8 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8.649996 ymin: 54.63326 xmax: -0.7246158 ymax: 60.86077
#> Geodetic CRS:  WGS 84
#>     spr22cd               spr22nm                       geometry
#> 1 S17000011 Highlands and Islands MULTIPOLYGON (((-5.572266 5...
#> 2 S17000012               Lothian MULTIPOLYGON (((-3.17174 55...
#> 3 S17000013 Mid Scotland and Fife MULTIPOLYGON (((-3.302574 5...
#> 4 S17000014   North East Scotland MULTIPOLYGON (((-2.531457 5...
#> 5 S17000015        South Scotland MULTIPOLYGON (((-4.085281 5...
#> 6 S17000018         West Scotland MULTIPOLYGON (((-5.118861 5...
#> 7 S17000019      Central Scotland MULTIPOLYGON (((-3.795887 5...
#> 8 S17000020               Glasgow MULTIPOLYGON (((-4.271141 5...
bounds("npark", within_names = "Bannau Brycheiniog")
#> Simple feature collection with 1 feature and 2 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -3.989641 ymin: 51.70209 xmax: -2.952245 ymax: 52.07827
#> Geodetic CRS:  WGS 84
#>   npark22cd          npark22nm                       geometry
#> 1 W18000001 Bannau Brycheiniog POLYGON ((-3.120246 52.0732...