Common procedure for joinable and not-necessarily-joinable lookup()s

Common procedure for joinable and not-necessarily-joinable lookup()s

Usage

common_lookup(
  lookup_level,
  within_level = NULL,
  within_names = NULL,
  within_codes = NULL,
  lookup_year = NULL,
  within_year = NULL,
  opts = boundr_options(),
  joinable = FALSE
)

Arguments

joinable

logical. Whether this query is joinable (TRUE) or forms part of a geospatial query (FALSE). In the latter case, this function will try to return a lookup table that contains the appropriate field for spatial data (that is, boundaries or centroids) to be joined onto.