Perform an API query and extract the returned IDs

For large queries, just return IDs that can then be batched for full queries.

Usage

return_result_ids(url, where, max_tries = 3, verbosity = 0, ...)

Arguments

url

The API URL to query

where

any geographical filters to be applied

max_tries

integer. 3 by default (for now). Passed to httr2::req_retry () and controls how many times to try to perform the request, should initial attempt(s) fail.

verbosity

integer. 0 by default. Passed to httr2::req_perform() and controls the verbosity of the printed output from performing the request. Can be any integer from 0 to 3, or NULL. See ?req_perform for more detail.

...

any arguments to be passed to opengeo_api_req()