Use an existing tibble as the basis for a spatial query

If you have a tibble such as those produced by lookup() - that is, there is a column of geographical ONS codes ending in 'cd' - simply use this table as the basis for retrieving the relevant boundaries or centroids.

Usage

add_geometry_to_table(
  tbl,
  geometry = c("boundaries", "centroids"),
  lookup_level = NULL,
  geo_code_field = NULL,
  opts = boundr_options()
)

add_geometry(
  tbl,
  geometry = c("boundaries", "centroids"),
  lookup_level = NULL,
  geo_code_field = NULL,
  opts = boundr_options()
)

Arguments

tbl

A tibble with a column for geographical codes. Table names ought to be in lower case. This function will use the lefthand-most column ending in 'cd' by default as the basis for retrieving boundary or point data.

geometry

character. Two options: "boundaries" (the default) and "centroids". By default, bounds() will return boundaries and centroids() will return centroids.

lookup_level

character. Lower level area code eg "lsoa", "wd", "lad".

geo_code_field

character The column name from tbl for which you wish to retrieve spatial data. By default, add_geometry_to_table() will use the lefthand-most column that ends in "cd".

opts

Should be set with the boundr_options() function. See ?boundr_options for detail on what can be set by the user, and on the default values.

Value

If successful, will return the initial table with an additional geometry column added. Duplicate rows will be removed. tibble::tibble(wd23cd = c("S13003001", "N08000520", "W05001522")) |> add_geometry_to_table(geo_code_field = "wd23cd")